+++ src/main/java/sisosolsol/greenfire/category/controller/CategoryController.java
@@ -5,7 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import sisosolsol.greenfire.category.model.dto.CategoryCreateDTO;
 import sisosolsol.greenfire.category.model.dto.CategoryUpdateDTO;
-import sisosolsol.greenfire.category.model.type.CategoryType;
+import sisosolsol.greenfire.common.enums.CategoryType;
 import sisosolsol.greenfire.category.service.CategoryService;
 
 import java.net.URI;
